Major Edits to hex-edit on MacOS:

+++ hiro/gtk/widget/hex-edit.cpp
@@ -94,7 +94,7 @@ auto pHexEdit::setAddress(u32 address) -> void {
   update();
 }
 
-auto pHexEdit::setBase(u16 base) -> void {
+auto pHexEdit::setBase(u8 base) -> void {
   setScroll();
   update();
 }
