+++ hiro/gtk/widget/hex-edit.hpp
@@ -7,7 +7,7 @@ struct pHexEdit : pWidget {
 
   auto focused() const -> bool override;
   auto setAddress(u32 address) -> void;
-  auto setBase(u16 base) -> void;
+  auto setBase(u8 base) -> void;
   auto setBackgroundColor(Color color) -> void;
   auto setColumns(u32 columns) -> void;
   auto setForegroundColor(Color color) -> void;
